fix: move arbitarysign into FirmaUtil and update test cases
